Remove two unused variables. Change a gint to gsize to avoid a crash when
2006-11-03  Jordi Mallach  <jordi@sindominio.net>

	* src/partyline.c: Remove two unused variables.
	* src/client.c (client_readmsg): Change a gint to gsize to avoid
	a crash when connecting on NetBSD/Sparc64. Thanks khorben for the
	info. Fixes: #338712.

+++ src/client.c
@@ -449,7 +449,7 @@ int client_sendmsg (char *str)
 
 int client_readmsg (gchar **str)
 {
-    gint bytes = 0;
+    gsize bytes = 0;
     gchar buf[1024];
     GError *error = NULL;
     gint i = 0;
